Move Documents74 namespace to Documents
