+++ psalm.xml.dist
@@ -15,7 +15,6 @@
         <directory name="lib/Doctrine/ODM/MongoDB" />
         <directory name="tests/Doctrine" />
         <directory name="tests/Documents" />
-        <directory name="tests/Documents74" />
         <ignoreFiles>
             <file name="lib/Doctrine/ODM/MongoDB/Aggregation/Stage/GraphLookup/Match.php" />
             <file name="lib/Doctrine/ODM/MongoDB/Aggregation/Stage/Match.php" />
