+++ tests/Doctrine/ODM/MongoDB/Tests/Mapping/AbstractMappingDriverTestCase.php
@@ -18,12 +18,11 @@ use Doctrine\ODM\MongoDB\Tests\BaseTestCase;
 use Doctrine\ODM\MongoDB\Types\Type;
 use Doctrine\Persistence\Mapping\Driver\MappingDriver;
 use Doctrine\Persistence\Reflection\EnumReflectionProperty;
-use Documents74\CustomCollection;
-use Documents74\UserTyped;
+use Documents\CustomCollection;
+use Documents\UserTyped;
 use Documents81\Card;
 use Documents81\Suit;
 use InvalidArgumentException;
-
 use function key;
 use function sprintf;
 use function strcmp;
