+++ tests/Doctrine/ODM/MongoDB/Tests/Mapping/ClassMetadataTest.php
@@ -22,13 +22,13 @@ use Documents\Album;
 use Documents\Bars\Bar;
 use Documents\CmsGroup;
 use Documents\CmsUser;
+use Documents\CustomCollection;
 use Documents\CustomRepository\Repository;
 use Documents\SpecialUser;
 use Documents\User;
 use Documents\UserName;
 use Documents\UserRepository;
-use Documents74\CustomCollection;
-use Documents74\UserTyped;
+use Documents\UserTyped;
 use Documents81\Card;
 use Documents81\Suit;
 use Documents81\SuitInt;
@@ -39,7 +39,6 @@ use ProxyManager\Proxy\GhostObjectInterface;
 use ReflectionClass;
 use ReflectionException;
 use stdClass;
-
 use function array_merge;
 use function MongoDB\BSON\fromJSON;
 use function MongoDB\BSON\toPHP;
