+++ tests/Documents/CustomCollection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Documents74;
+namespace Documents;
 
 use Doctrine\Common\Collections\ArrayCollection;
