+++ tests/Documents/TypedDocument.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Documents74;
+namespace Documents;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
